GH 45c & 46, PP 34, RP 21, OW 123e, VTR 20, MCI 134, TG 5a, HY 11a, WH 130a, SP 270 + Misc
Fix build errors with ET with reverted GPS DLL.

+++ NRaasTagger/TaggerSpace/Options/MetaAutonomyOptions/HourCloseOption.cs
@@ -42,7 +42,7 @@ namespace NRaas.TaggerSpace.Options.MetaAutonomyOptions
             {
                 MetaAutonomySettingKey key = Tagger.Settings.GetMASettings(mType);
                 key.mHourClose = value;
-                Tagger.Settings.AddOrUpdateMASettings(mType, key);
+                Tagger.Settings.AddOrUpdateMASettings((uint) mType, key);
             }
         }
